Make local partition inspection explicit
A position's own type and rules were read by the walk, in pieces: the classes its
type states, the range its rules leave, the lines they draw. Whether the local
reading had run out was then a caller noticing that two lists were both empty —
and what follows from that noticing is the rest of the derivation, since only an
exhausted local reading licenses asking what is under the position. Anything that
read one list and went on would have concluded it from half the evidence.

`LocalInspection` is that answer as a value. Both producers are asked to build
one, so `Exhausted` is not "there were no classes": it is that both were asked
and neither had anything. `Evidence` cannot be built empty, and the two are the
whole of what a local reading comes to.

What the reading saw is apart from what it concluded. `LocalReading` — the term
the position is measured at, what its rules leave it, which of them went unread —
is one value held by either answer, because one reading produced both. Copied
into each, the day someone works out the admissible range differently for a
position that has classes is the day the compiler contradicts itself about where
the same values stop.

`CutEvidence` tells no lines from lines. Whether a row at an edge may turn out
unwritable is a fact about edges, so a position with none cannot carry the
answer: `uncertain` was a boolean beside a possibly-empty list, true or false by
whoever set it.

Three readings come out of the orchestration and become what they are. `TypeBounds`
is what a type's own rules leave a value between, read once for the two things
that need it — where a position is divided, and what values can be built to stand
at it. `PartitionClasses` is the class producer, which the position's reading, a
witness of a type and the search for a representative all now ask by name. The
javadoc that says an empty answer is this producer having no evidence — never
that the position has no classes — belongs with it rather than in a walk's
private helper.

The precedence is then the arm the code is written in. A position its own type
answered for is not descended into; only where the local reading is exhausted is
the structural question asked, and only `Children` takes the position away.

A position with both local evidence and children is intended and unreachable
today: only products have children, and a product carries neither classes nor
cuts. Which of the two wins is written down as a decision anyway — a field-level
partition or an invariant over a whole product would inhabit it, and an
implementation that descended because the position is structural would then lose
evidence it had already read. The test says as much rather than building one out
of a hand-made shape, which would fix the implementation's product space instead
of the language's.

No report changes: the corpus output is identical, which is what a move of this
size has to show.
